# schema-stub/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "IdentityChain Agent REST API"
  version: "0.1.6"
  title: "IdentityChain API"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Bearer: []
paths:
...
  /api/wallet:
    get:
      tags:
      - "wallet"
      summary: "List all wallets of user"
      description: "List all wallets of user"
      operationId: "apiWalletGET"
      parameters: []
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Wallet"
    post:
      tags:
      - "wallet"
      summary: "Create a new Wallet (optionally with given name and settings)"
      description: "Create a new Wallet (optionally with given name and settings)"
      operationId: "apiWalletPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/wallet_post"
      responses:
        201:
          description: ""
      x-swagger-router-controller: "Wallet"
  /api/wallet/{wallet}:
    get:
      tags:
      - "wallet"
      summary: "Retrieve a Wallet"
      description: "Retrieve a Wallet"
      operationId: "apiWalletWalletGET"
      parameters:
      - name: "wallet"
        in: "path"
        description: "A unique id value (or \"default\") identifying this wallet."
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Wallet"
    delete:
      tags:
      - "wallet"
      summary: "Delete a Wallet"
      description: "Delete a Wallet"
      operationId: "apiWalletWalletDELETE"
      parameters:
      - name: "wallet"
        in: "path"
        description: "A unique id value (or \"default\") identifying this wallet."
        required: true
        type: "string"
      responses:
        204:
          description: ""
      x-swagger-router-controller: "Wallet"
...
securityDefinitions:
  Bearer:
    description: "JWT Authorization header using the Bearer scheme.\nExample: 'Authorization:\
      \ Bearer {token}'\n\nPlease add Bearer in front of the JWT token:\n> Bearer\
      \ xxxxx.yyyyy.zzzzz\n"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  wallet_post:
    type: "object"
    required:
    - "credentials"
    properties:
      name:
        type: "string"
        description: "(Optional) Wallet name, must be globally unique. Automatically\
          \ generated if none is provided."
      credentials:
        $ref: "#/definitions/wallet_post_credentials"
      seed:
        type: "string"
        description: "(Optional) Seed to use for initial did creation."
...
  wallet_post_credentials:
    required:
    - "key"
    properties:
      key:
        type: "string"
        description: "Passphrase used to derive wallet master key"
    description: "Wallet Credentials JSON. Supported keys vary by wallet type. A default\
      \ config will be used if none is provided."
...